feat(frontend): Projects moves out — entity layer in entities, table in settings-ui
The service consolidates onto one transport: OSS ran GETs through its own fetch
client and writes through axios, which a package cannot reach and which had no
behavioural reason. All five calls now use @agenta/shared's axios and its auth
interceptor; fetchAllProjects still answers [] on 401, as its callers expect.

The create/rename/delete dialogs stay with the host: they are antd Form with
validation rules, and settings-ui is antd-free by contract. The page owns the
table, the toolbar and the mutations and takes the dialogs as slots.

/m's own fetchProjects still exists and is the better implementation — it retries
after a session refresh. Folding that in would change desktop auth behaviour, so
it stays a follow-up rather than riding along with a table move.
