+++ web/oss/src/services/project/types.ts
@@ -1,11 +1,1 @@
-export interface ProjectsResponse {
-    project_id: string
-    project_name: string
-    workspace_id?: string | null
-    workspace_name?: string | null
-    organization_id?: string | null
-    organization_name?: string | null
-    user_role?: string | null
-    is_demo?: boolean | null
-    is_default_project?: boolean
-}
+export * from "@agenta/entities/project"
