+++ web/packages/agenta-settings-ui/package.json
@@ -24,16 +24,18 @@
         "clsx": "^2.1.1"
     },
     "peerDependencies": {
+        "@tanstack/react-query": ">=5.0.0",
         "jotai": ">=2.0.0",
-        "react": ">=18.0.0",
-        "next": ">=14.0.0"
+        "next": ">=14.0.0",
+        "react": ">=18.0.0"
     },
     "devDependencies": {
+        "@tanstack/react-query": "5.100.9",
         "@types/node": "^20.19.20",
         "@types/react": "^19.0.10",
         "jotai": "^2.15.0",
+        "next": "15.5.18",
         "react": "^19.0.0",
-        "typescript": "^5.9.3",
-        "next": "15.5.18"
+        "typescript": "^5.9.3"
     }
 }
