+++ web/packages/agenta-settings-ui/src/index.ts
@@ -13,3 +13,4 @@ export {ApiKeysPage, type ApiKeysPageProps} from "./ApiKeysPage"
 export {SecretProviderTable, type SecretProviderTableProps, type ProviderDialogState} from "./secrets/SecretProviderTable"
 export {NamedSecretTable, type NamedSecretTableProps} from "./secrets/NamedSecretTable"
 export {WebhooksPage, type WebhooksPageProps} from "./webhooks/WebhooksPage"
+export {ProjectsPage, type ProjectsPageProps, type ProjectDialogState} from "./projects/ProjectsPage"
